Complete form validation function using API
Upon events of input and form submit, use validation API to validate
required fields and range of page number.

If found invalid, it set custom validation message then report validity
to the user.

Replace `clearFormInputs` with built-in `form.reset` method to reset
inputs of form, it also resets the validity state of inputs.